requirements.txt
# Common
streamlit

# Dependencies
pandas-profiling
requests

# Components
streamlit-ace
streamlit-elements
streamlit-pandas-profiling
